Noveck suggested:
"I think we need more clarity about whether the format cannot be
described in this document or whether it could but is not present
because it is unnecessary."

I've chosen B, because IMA metadata is never interpreted by either
local or remote file systems. The Introduction has been updated to
reflect this choice.